Move geff comparison utility into src from tests
